openSNP data

openSNP data

opensnp

Format

The data set contains 23 variables of 784 individuals. Twenty-one of these variables represent minor allele counts of SNPs, where the name is a combination of SNP identifier and the minor allele.

id

Personal identifier

height

Human height in cm

rs1985272_C

Minor allele counts for SNP identifier rs1985272, minor allele: C.

...

rs10489858_C

Minor allele counts for SNP identifier rs10489858, minor allele: C.

Source

Data : CrowdAI / OpenSNP - height prediction challenge by Oliver Naret, https://zenodo.org/record/1442755#.W9LfWSexVOE Manuscript: openSNP: https://opensnp.org/ and Greshake B, Bayer PE, Rausch H, Reda J (2014) openSNP<U+2013>A Crowdsourced Web Resource for Personal Genomics. PLOS ONE 9(3): e89204. https://doi.org/10.1371/journal.pone.0089204

Details

Minor allele count per person is the number of minor allele an individual carries, possible values are 0, 1 and 2. For example, for variable rs1985272_C the minor allele of SNP id rs1985272 is C. In this case, an entry of 0 means that this individual has no minor allele C, 1 means this individual has one minor allele C, and 2 means this person has two copies of the minor allele C.